Added `...` to `init.parms()` generic
This allows the function to be more flexible (https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Generic-functions-and-methods
itself says that "there need to be good arguments not to include a … argument"). More importantly for issue #88 and this commit in particular, this allows the `trunc_binomial` method to check for the presence of `nsize`.