Improved flexibility of output for *binomial (#87)
Estimation of the original parameters through `empiricalParameters()` in those cases requires the user to inform the size (Binomial and Negative Binomial, named r in the latter) and the number of failures (k, for nbinom only). This is useful if the user is calling the function directly. For ML estimation, it's enough to use `mean(y)`.